server/requirements.txt
fastapi
uvicorn
gunicorn
python-jose
pydantic[email]
pydantic[dotenv]
loguru
passlib
cryptography
pycryptodome
pymysql
pymongo
redis
amqpstorm
SQLAlchemy
bcrypt
tenacity
celery
requests
psycopg2
python-multipart
